+++ lib/aspera/cli/plugins/config.rb
@@ -747,7 +747,7 @@ module Aspera
             return Main.result_status("Installed #{n} version #{v}")
           when :spec
             return Main.result_object_list(
-              Transfer::SpecDoc.man_table(formatter, cli: false),
+              Transfer::SpecDoc.man_table(Formatter, cli: false),
               fields: Transfer::SpecDoc::TABLE_COLUMNS.map(&:to_s)
             )
           when :schema