+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     * `transferd`: Fixed discrepancies in transfer spec resume policies.
     * `desktop`: Fixed discrepancies in transfer spec resume policies.
     * `format`: Value list now properly display column header.
+    * `select`: Filter is now done on values before enhanced display in table mode.
   * Breaking Changes:
     * `ts` : Default transfer spec includes `resume_policy=sparse_csum`
     * `ssh_options` : Now additive option, like `ts`.